# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:50182","description":"Generated server url"}],"paths":{"/hotels/reserva":{"post":{"tags":["hotel-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationWriteDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/hotels/reserva/{id_reserva}":{"get":{"tags":["hotel-controller"],"operationId":"getReservation","parameters":[{"name":"id_reserva","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}}},"components":{"schemas":{"ReservationWriteDTO":{"type":"object","properties":{"id_hotel":{"type":"integer","format":"int64"},"id_cliente":{"type":"string"},"tipo_habitacion":{"type":"string","enum":["SIMPLE","DOUBLE","TRIPLE"]},"fecha_ingreso":{"type":"string","format":"date"},"fecha_salida":{"type":"string","format":"date"},"medio_pago":{"type":"string","enum":["EFECTIVO","TRANSFERENCIA"]}}},"ReservationDTO":{"type":"object","properties":{"id_reserva":{"type":"integer","format":"int64"},"id_cliente":{"type":"string"},"id_hotel":{"type":"integer","format":"int64"},"tipo_habitacion":{"type":"string","enum":["SIMPLE","DOUBLE","TRIPLE"]},"fecha_ingreso":{"type":"string","format":"date"},"fecha_salida":{"type":"string","format":"date"},"estado_reserva":{"type":"string","enum":["EXITOSA","CANCELADA","PENDIENTE"]},"precio":{"type":"number"},"medio_pago":{"type":"string","enum":["EFECTIVO","TRANSFERENCIA"]}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}